+++ internal/key/templates.go
@@ -44,6 +44,12 @@ const ClusterCAPACRsTemplate = `
 {{ .AWSMachineTemplateCR -}}
 ---
 {{ .AWSClusterRoleIdentityCR -}}
+---
+{{ .BastionBootstrapSecret -}}
+---
+{{ .BastionMachineDeploymentCR -}}
+---
+{{ .BastionAWSMachineTemplateCR -}}
 `
 
 const MachineDeploymentCRsTemplate = `
@@ -70,4 +76,90 @@ const MachinePoolAzureCRsTemplate = `
 {{ .MachinePoolCR -}}
 ---
 {{ .SparkCR -}}
+`
+
+const BastionIgnitionTemplate = `{
+   "ignition":{
+      "version":"2.2.0"
+   },
+   "passwd":{
+      "users":[
+         {
+            "name":"giantswarm",
+            "sshAuthorizedKeys":[
+               "ssh-rsa AAAABEKf fake@giantswarm"
+            ],
+            "groups":[
+               "sudo",
+               "docker"
+            ],
+            "shell":"/bin/bash",
+            "uid":1000
+         }
+      ]
+   },
+   "storage":{
+      "files":[
+         {
+            "path":"/etc/hostname",
+            "filesystem":"root",
+            "mode": 420,
+            "contents":{
+               "source":"data:,%s-bastion"
+            }
+         },
+         {
+            "path":"/etc/ssh/sshd_config",
+            "filesystem":"root",
+            "mode": 420,
+            "contents":{
+               "source":"data:text/plain;charset=utf-8;base64,%s"
+            }
+         },
+         {
+            "path":"/etc/ssh/trusted-user-ca-keys.pem",
+            "filesystem":"root",
+            "mode": 420,
+            "contents":{
+               "source":"data:text/plain;charset=utf-8;base64,%s"
+            }
+         }
+      ]
+   }
+}`
+
+const ubuntuSudoersConfig = "giantswarm ALL=(ALL:ALL) NOPASSWD: ALL"
+
+const bastionSSHDConfig = `# Use most defaults for sshd configuration.
+Subsystem sftp internal-sftp
+ClientAliveInterval 180
+UseDNS no
+UsePAM yes
+PrintLastLog no # handled by PAM
+PrintMotd no # handled by PAM
+# Non defaults (#100)
+ClientAliveCountMax 2
+PasswordAuthentication no
+TrustedUserCAKeys /etc/ssh/trusted-user-ca-keys.pem
+MaxAuthTries 5
+LoginGraceTime 60
+AllowTcpForwarding yes
+AllowAgentForwarding yes
+CASignatureAlgorithms ecdsa-sha2-nistp256,ecdsa-sha2-nistp384,ecdsa-sha2-nistp521,ssh-ed25519,rsa-sha2-512,rsa-sha2-256,ssh-rsa`
+
+const nodeSSHDConfig = `# Use most defaults for sshd configuration.
+Subsystem sftp internal-sftp
+ClientAliveInterval 180
+UseDNS no
+UsePAM yes
+PrintLastLog no # handled by PAM
+PrintMotd no # handled by PAM
+# Non defaults (#100)
+ClientAliveCountMax 2
+PasswordAuthentication no
+TrustedUserCAKeys /etc/ssh/trusted-user-ca-keys.pem
+MaxAuthTries 5
+LoginGraceTime 60
+AllowTcpForwarding no
+AllowAgentForwarding no
 `